requirements.txt
sanic
sanic-ext
orjson
python-dotenv
openai
streamlit
pandas
python-docx
PyPDF2
fpdf
spacy
langchain
langchain-community
